aibackend/requirements.txt
fastapi
uvicorn
torch
torchvision
opencv-python
tensorflow-macos
numpy
